extract RequestHandler class
